sign up func moved to app
