+++ src/components/Cart.js
@@ -1,6 +1,15 @@
+import { useEffect, useState } from 'react';
+
 
 function Cart(props) {
-    const { userProductUnits } = props
+    const { userProductUnits, fetchUserProductUnits, user } = props
+    
+    useEffect(() => {
+        fetchUserProductUnits(user.userID)
+        
+        // setCartItems(JSON.parse(localStorage.cartItems))
+        // console.log(JSON.parse(localStorage.getItem("cartItems")))
+      }, []);
 
     return (
         <div>
