+++ src/components/register.js
@@ -5,44 +5,31 @@ import { useState } from "react";
 
 // const {createUser} = require('../../db/users');
 
-const createUser = async (username, password) => {
-    console.log('This is the createUser func');
+// const createUser = async (username, password) => {
+//     console.log('This is the createUser func');
 
-    console.log("new username and password", username, password)
+//     console.log("new username and password", username, password)
     
-    const response = await fetch('http://localhost:4000/api/users/signup', {
-        
+//     const response = await fetch('http://localhost:4000/api/users/signup', {
 
-//     const response = await fetch('https:localhost:4000/api/signup', {
+//         method: "POST",
+//         headers: {
+//             'Content-Type' : 'application/json',
+//         },
 
-        method: "POST",
-        headers: {
-            'Content-Type' : 'application/json',
-        },
-
-        body: JSON.stringify({
-            username: username,
-            password: password
-        }),
-        mode: "cors",
-    });
-    console.log("this is the response", response)
-
-//         body: {
+//         body: JSON.stringify({
 //             username: username,
 //             password: password
-//         }
+//         }),
+//         mode: "cors",
 //     });
-
-    // if (response) {
-    //     const {token } = await response.json();
-    //     localStorage.setItem("token", token)
-    // }
-    return response;
-}
+//     console.log("this is the response", response)
+//     return response;
+// }
 
 
-const Register = () => {
+const Register = (props) => {
+    const {createUser} = props
     const [username, setUsername] = useState('');
     const [password, setPassword] = useState('');
